refactor: renamed component name for layer ordering feature because it is confused with other components
